requirements.txt
Flask==3.0.3
plyer==2.1.0
python-dotenv==1.0.1
tweepy==4.14.0